Remove singletons, fix remaining test fails
