+++ lib/runtime/AllocationTracking.cpp
@@ -4,13 +4,14 @@
 
 #include "AllocationTracking.h"
 
-#include "CallbackInterface.h"
-#include "RuntimeData.h"
-#include "support/Logger.h"
-#include "AccessCounter.h"
 #include "AccessCountPrinter.h"
+#include "AccessCounter.h"
+#include "CallbackInterface.h"
+#include "Runtime.h"
+#include "RuntimeData.h"
 #include "RuntimeUtils.h"
 #include "TypeResolution.h"
+#include "support/Logger.h"
 
 #include <algorithm>
 #include <cassert>
@@ -34,87 +35,8 @@ using namespace btree;
 
 namespace typeart {
 
-enum class AllocState : unsigned {
-  NO_INIT      = 1 << 0,
-  OK           = 1 << 1,
-  ADDR_SKIPPED = 1 << 2,
-  NULL_PTR     = 1 << 3,
-  ZERO_COUNT   = 1 << 4,
-  NULL_ZERO    = 1 << 5,
-  ADDR_REUSE   = 1 << 6,
-  UNKNOWN_ID   = 1 << 7
-};
-
-enum class FreeState : unsigned {
-  NO_INIT      = 1 << 0,
-  OK           = 1 << 1,
-  ADDR_SKIPPED = 1 << 2,
-  NULL_PTR     = 1 << 3,
-  UNREG_ADDR   = 1 << 4,
-};
-
-struct AllocationData {
-  AllocationData() {
-    stackVars.reserve(RuntimeT::StackReserve);
-    debug::printTraceStart();
-  }
-
-  RuntimeT::PointerMap allocTypes;
-  RuntimeT::Stack stackVars;
-};
-
-namespace detail {
-template <class...>
-constexpr std::false_type always_false{};
-}  // namespace detail
-
-template <typename Enum>
-inline Enum operator|(Enum lhs, Enum rhs) {
-  if constexpr (std::is_enum_v<Enum> &&
-                (std::is_same_v<Enum, AllocState> || std::is_same_v<Enum, FreeState>)) {
-    using enum_t = typename std::underlying_type<Enum>::type;
-    return static_cast<Enum>(static_cast<enum_t>(lhs) | static_cast<enum_t>(rhs));
-  } else {
-    static_assert(detail::always_false<Enum>);
-  }
-}
-template <typename Enum>
-inline void operator|=(Enum& lhs, Enum rhs) {
-  if constexpr (std::is_enum_v<Enum> &&
-                (std::is_same_v<Enum, AllocState> || std::is_same_v<Enum, FreeState>)) {
-    lhs = lhs | rhs;
-  } else {
-    static_assert(detail::always_false<Enum>);
-  }
-}
-
-template <typename Enum>
-inline Enum operator&(Enum lhs, Enum rhs) {
-  if constexpr (std::is_enum_v<Enum> && std::is_same_v<Enum, AllocState>) {
-    using enum_t = typename std::underlying_type<Enum>::type;
-    return static_cast<Enum>(static_cast<enum_t>(lhs) & static_cast<enum_t>(rhs));
-  } else {
-    static_assert(detail::always_false<Enum>);
-  }
-}
-
-template <typename Enum>
-inline typename std::underlying_type<Enum>::type operator==(Enum lhs, Enum rhs) {
-  if constexpr (std::is_enum_v<Enum> && std::is_same_v<Enum, AllocState>) {
-    using enum_t = typename std::underlying_type<Enum>::type;
-    return static_cast<enum_t>(lhs) & static_cast<enum_t>(rhs);
-  } else {
-    static_assert(detail::always_false<Enum>);
-  }
-}
-
 using namespace debug;
 
-namespace {
-// Global allocation data
-AllocationData allocData;
-
-// TODO: Add mutex
 /**
  * Ensures that memory tracking functions are not called from within the runtime.
  * TODO: Problematic with respect to future thread safety considerations
@@ -122,9 +44,9 @@ AllocationData allocData;
  */
 static bool typeart_rt_scope{false};
 
-AllocState doAlloc(const void* addr, int typeId, size_t count, const void* retAddr) {
+AllocState AllocationTracker::doAlloc(const void* addr, int typeId, size_t count, const void* retAddr) {
   AllocState status = AllocState::NO_INIT;
-  if (unlikely(!TypeResolution::get().isValidType(typeId))) {
+  if (unlikely(!runtime.typeResolution.isValidType(typeId))) {
     status |= AllocState::UNKNOWN_ID;
     LOG_ERROR("Allocation of unknown type " << toString(addr, typeId, count, retAddr));
   }
@@ -133,24 +55,24 @@ AllocState doAlloc(const void* addr, int typeId, size_t count, const void* retAd
   // In the second case, the allocation is tracked anyway so that onFree() does not report an error.
   // On the other hand, an allocation on address 0x0 with size > 0 is an actual error.
   if (unlikely(count == 0)) {
-    Recorder::get().incZeroLengthAddr();
+    runtime.recorder.incZeroLengthAddr();
     status |= AllocState::ZERO_COUNT;
     LOG_WARNING("Zero-size allocation " << toString(addr, typeId, count, retAddr));
     if (addr == nullptr) {
-      Recorder::get().incZeroLengthAndNullAddr();
+      runtime.recorder.incZeroLengthAndNullAddr();
       LOG_ERROR("Zero-size and nullptr allocation " << toString(addr, typeId, count, retAddr));
       return status | AllocState::NULL_ZERO | AllocState::ADDR_SKIPPED;
     }
   } else if (unlikely(addr == nullptr)) {
-    Recorder::get().incNullAddr();
+    runtime.recorder.incNullAddr();
     LOG_ERROR("Zero-size allocation " << toString(addr, typeId, count, retAddr));
     return status | AllocState::NULL_PTR | AllocState::ADDR_SKIPPED;
   }
 
-  auto& def = allocData.allocTypes[addr];
+  auto& def = runtime.allocTracker.allocTypes[addr];
 
   if (unlikely(def.typeId != -1)) {
-    typeart::Recorder::get().incAddrReuse();
+    typeart::runtime.recorder.incAddrReuse();
     status |= AllocState::ADDR_REUSE;
     LOG_WARNING("Pointer already in map " << toString(addr, typeId, count, retAddr));
     LOG_WARNING("Overriden data in map " << toString(addr, def));
@@ -163,39 +85,40 @@ AllocState doAlloc(const void* addr, int typeId, size_t count, const void* retAd
   return status | AllocState::OK;
 }
 
-void onAlloc(const void* addr, int typeId, size_t count, const void* retAddr) {
+void AllocationTracker::onAlloc(const void* addr, int typeId, size_t count, const void* retAddr) {
   const auto status = doAlloc(addr, typeId, count, retAddr);
   if (status != AllocState::ADDR_SKIPPED) {
-    typeart::Recorder::get().incHeapAlloc(typeId, count);
+    typeart::runtime.recorder.incHeapAlloc(typeId, count);
   }
   LOG_TRACE("Alloc " << toString(addr, typeId, count, retAddr) << " " << 'H');
 }
 
-void onAllocStack(const void* addr, int typeId, size_t count, const void* retAddr) {
+void AllocationTracker::onAllocStack(const void* addr, int typeId, size_t count, const void* retAddr) {
   const auto status = doAlloc(addr, typeId, count, retAddr);
   if (status != AllocState::ADDR_SKIPPED) {
-    allocData.stackVars.push_back(addr);
-    typeart::Recorder::get().incStackAlloc(typeId, count);
+    runtime.allocTracker.stackVars.push_back(addr);
+    typeart::runtime.recorder.incStackAlloc(typeId, count);
   }
   LOG_TRACE("Alloc " << toString(addr, typeId, count, retAddr) << " " << 'S');
 }
 
-void onAllocGlobal(const void* addr, int typeId, size_t count, const void* retAddr) {
+void AllocationTracker::onAllocGlobal(const void* addr, int typeId, size_t count, const void* retAddr) {
   const auto status = doAlloc(addr, typeId, count, retAddr);
   if (status != AllocState::ADDR_SKIPPED) {
-    typeart::Recorder::get().incGlobalAlloc(typeId, count);
+    typeart::runtime.recorder.incGlobalAlloc(typeId, count);
   }
   LOG_TRACE("Alloc " << toString(addr, typeId, count, retAddr) << " " << 'G');
 }
 
 template <bool stack>
-FreeState doFree(const void* addr, const void* retAddr) {
+FreeState AllocationTracker::doFree(const void* addr, const void* retAddr) {
   if (unlikely(addr == nullptr)) {
     LOG_ERROR("Free on nullptr "
-                  << "(" << retAddr << ")");
+              << "(" << retAddr << ")");
     return FreeState::ADDR_SKIPPED | FreeState::NULL_PTR;
   }
 
+  auto& allocData = runtime.allocTracker;
 
   const auto it = allocData.allocTypes.find(addr);
 
@@ -206,9 +129,9 @@ FreeState doFree(const void* addr, const void* retAddr) {
       const auto typeId = it->second.typeId;
       const auto count  = it->second.count;
       if (stack) {
-        Recorder::get().incStackFree(typeId, count);
+        runtime.recorder.incStackFree(typeId, count);
       } else {
-        Recorder::get().incHeapFree(typeId, count);
+        runtime.recorder.incHeapFree(typeId, count);
       }
     }
 
@@ -221,15 +144,15 @@ FreeState doFree(const void* addr, const void* retAddr) {
   return FreeState::OK;
 }
 
-void onFreeHeap(const void* addr, const void* retAddr) {
+void AllocationTracker::onFreeHeap(const void* addr, const void* retAddr) {
   const auto status = doFree<false>(addr, retAddr);
   if (FreeState::OK == status) {
-    typeart::Recorder::get().decHeapAlloc();
+    typeart::runtime.recorder.decHeapAlloc();
   }
 }
 
-void onLeaveScope(int alloca_count, const void* retAddr) {
-  auto& stackVars = allocData.stackVars;
+void AllocationTracker::onLeaveScope(int alloca_count, const void* retAddr) {
+  auto& stackVars = runtime.allocTracker.stackVars;
 
   if (unlikely(alloca_count > stackVars.size())) {
     LOG_ERROR("Stack is smaller than requested de-allocation count. alloca_count: " << alloca_count
@@ -240,16 +163,14 @@ void onLeaveScope(int alloca_count, const void* retAddr) {
   const auto cend      = stackVars.cend();
   const auto start_pos = (cend - alloca_count);
   LOG_TRACE("Freeing stack (" << alloca_count << ")  " << std::distance(start_pos, stackVars.cend()))
-  std::for_each(start_pos, cend, [&retAddr](const void* addr) { doFree<true>(addr, retAddr); });
+  std::for_each(start_pos, cend, [this, &retAddr](const void* addr) { doFree<true>(addr, retAddr); });
   stackVars.erase(start_pos, cend);
-  typeart::Recorder::get().decStackAlloc(alloca_count);
+  typeart::runtime.recorder.decStackAlloc(alloca_count);
   LOG_TRACE("Stack after free: " << stackVars.size());
 }
 
-}  // namespace
-
-llvm::Optional<RuntimeT::MapEntry> findBaseAlloc(const void* addr) {
-  auto& allocs = allocData.allocTypes;
+llvm::Optional<RuntimeT::MapEntry> AllocationTracker::findBaseAlloc(const void* addr) {
+  auto& allocs = allocTypes;
   if (allocs.empty() || addr < allocs.begin()->first) {
     return llvm::None;
   }
@@ -273,34 +194,34 @@ llvm::Optional<RuntimeT::MapEntry> findBaseAlloc(const void* addr) {
 void __typeart_alloc(const void* addr, int typeId, size_t count) {
   RUNTIME_GUARD_BEGIN;
   const void* retAddr = __builtin_return_address(0);
-  typeart::onAlloc(addr, typeId, count, retAddr);
+  typeart::runtime.allocTracker.onAlloc(addr, typeId, count, retAddr);
   RUNTIME_GUARD_END;
 }
 
 void __typeart_alloc_stack(const void* addr, int typeId, size_t count) {
   RUNTIME_GUARD_BEGIN;
   const void* retAddr = __builtin_return_address(0);
-  typeart::onAllocStack(addr, typeId, count, retAddr);
+  typeart::runtime.allocTracker.onAllocStack(addr, typeId, count, retAddr);
   RUNTIME_GUARD_END;
 }
 
 void __typeart_alloc_global(const void* addr, int typeId, size_t count) {
   RUNTIME_GUARD_BEGIN;
   const void* retAddr = __builtin_return_address(0);
-  typeart::onAllocGlobal(addr, typeId, count, retAddr);
+  typeart::runtime.allocTracker.onAllocGlobal(addr, typeId, count, retAddr);
   RUNTIME_GUARD_END;
 }
 
 void __typeart_free(const void* addr) {
   RUNTIME_GUARD_BEGIN;
   const void* retAddr = __builtin_return_address(0);
-  typeart::onFreeHeap(addr, retAddr);
+  typeart::runtime.allocTracker.onFreeHeap(addr, retAddr);
   RUNTIME_GUARD_END;
 }
 
 void __typeart_leave_scope(int alloca_count) {
   RUNTIME_GUARD_BEGIN;
   const void* retAddr = __builtin_return_address(0);
-  typeart::onLeaveScope(alloca_count, retAddr);
+  typeart::runtime.allocTracker.onLeaveScope(alloca_count, retAddr);
   RUNTIME_GUARD_END;
 }
