+++ lib/runtime/TypeResolution.h
@@ -5,25 +5,19 @@
 #ifndef TYPEART_TYPERESOLUTION_H
 #define TYPEART_TYPERESOLUTION_H
 
+#include "RuntimeData.h"
 #include "RuntimeInterface.h"
 #include "TypeDB.h"
 
 namespace typeart {
 
-
 class TypeResolution {
   TypeDB typeDB;
-
-  TypeResolution();
 
  public:
   using TypeArtStatus = typeart_status;
 
-  static TypeResolution& get() {
-    // Note: This is guaranteed to be thread-safe
-    static TypeResolution instance;
-    return instance;
-  }
+  TypeResolution();
 
   size_t getMemberIndex(typeart_struct_layout structInfo, size_t offset) const;
 
@@ -36,12 +30,12 @@ class TypeResolution {
   TypeArtStatus getTypeInfoInternal(const void* baseAddr, size_t offset, const StructTypeInfo& containerInfo, int* type,
                                     size_t* count) const;
 
-  TypeArtStatus getTypeInfo(const void* addr, int* type, size_t* count) const;
+  TypeArtStatus getTypeInfo(const void* addr, RuntimeT::MapEntry allocInfo, int* type, size_t* count) const;
 
-  TypeArtStatus getContainingTypeInfo(const void* addr, int* type, size_t* count, const void** baseAddress,
-                                      size_t* offset) const;
+  TypeArtStatus getContainingTypeInfo(const void* addr, RuntimeT::MapEntry allocInfo, int* type, size_t* count,
+                                      const void** baseAddress, size_t* offset) const;
 
-  TypeArtStatus getBuiltinInfo(const void* addr, typeart::BuiltinType* type) const;
+  TypeArtStatus getBuiltinInfo(const void* addr, RuntimeT::MapEntry alloc, typeart::BuiltinType* type) const;
 
   TypeArtStatus getStructInfo(int id, const StructTypeInfo** structInfo) const;
 
@@ -52,6 +46,6 @@ class TypeResolution {
   bool isValidType(int id) const;
 };
 
-}
+}  // namespace typeart
 
 #endif  // TYPEART_TYPERESOLUTION_H
